web/requirements.txt
Flask==3.1.0
flask-mysql-connector
requests